HETS-1294 update behaviour of rental rate period field
The dropdown for rate period on the included rates and attachments dialog has been changed into an uneditable label that displays the equipment rate period.

The separate checkbox to indicate a "Set" period on the as-needed rates and attachments dialog has been removed and an additional option has been added to the rate period dropdown for it. This change does not affect the back-end or database in relation to how the RatePeriod and Set properties are used. Choosing the "Set" option in the dropdown changes the Set property of the rate to true and the RatePeriod to match the equipment rate period.